Promote 3.1.0 and organize facility creative tabs
