+++ build.gradle
@@ -3,7 +3,7 @@ plugins {
     id 'net.minecraftforge.gradle' version '[6.0,6.2)'
 }
 
-version = '3.0.8'
+version = '3.1.0'
 group = 'com.bl4ues.scpadditions'
 archivesBaseName = 'scp_additions'
