+++ src/main/resources/assets/scp_additions/models/item/facility_blocks_tab_icon.json
@@ -1,0 +1,3 @@
+{
+  "parent": "builtin/entity"
+}
