+++ src/main/resources/assets/scp_additions/models/item/facility_tab_divider.json
@@ -1,0 +1,6 @@
+{
+  "parent": "minecraft:item/generated",
+  "textures": {
+    "layer0": "scp_additions:item/facility_tab_divider"
+  }
+}
