+++ src/main/resources/assets/scp_additions/models/item/scps_tab_icon.json
@@ -1,0 +1,3 @@
+{
+  "parent": "builtin/entity"
+}
